+++ SmartOffice/src/main/java/com/artifex/sonui/editor/SlideShowConductorAnimations/WheelFadeAnimation.java
@@ -2,6 +2,7 @@ package com.artifex.sonui.editor.SlideShowConductorAnimations;
 
 import android.graphics.Path;
 import android.graphics.RectF;
+
 import com.artifex.sonui.editor.SlideShowConductorView;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
 
@@ -26,7 +27,7 @@ public class WheelFadeAnimation extends ShapeFadeAnimation {
         }
         int i = this.mHeight;
         int i2 = this.mWidth;
-        int sqrt = (int) Math.sqrt((double) ((i2 * i2) + (i * i)));
+        int sqrt = (int) Math.sqrt((i2 * i2) + (i * i));
         int i3 = this.mWidth;
         int i4 = this.mHeight;
         RectF rectF = new RectF((float) ((i3 / 2) - sqrt), (float) ((i4 / 2) - sqrt), (float) ((i3 / 2) + sqrt), (float) ((i4 / 2) + sqrt));
