+++ SmartOffice/src/main/java/com/artifex/sonui/editor/animations/WedgeAnimation.java
@@ -13,7 +13,7 @@ public class WedgeAnimation extends ShapeAnimation {
         float f2 = f * 180.0f;
         int i = this.mHeight;
         int i2 = this.mWidth;
-        int sqrt = (int) Math.sqrt((double) ((i2 * i2) + (i * i)));
+        int sqrt = (int) Math.sqrt((i2 * i2) + (i * i));
         int i3 = this.mWidth;
         int i4 = this.mHeight;
         RectF rectF = new RectF((float) ((i3 / 2) - sqrt), (float) ((i4 / 2) - sqrt), (float) ((i3 / 2) + sqrt), (float) ((i4 / 2) + sqrt));
